Split launcher chain args test
