+++ .pm/tasks/task_d018ce74bc994be0a0feb2cc47a14630.execution.md
@@ -88,3 +88,24 @@ Example:
 - Expected Result: `.pm` task and execution log satisfy repo lint.
 - Actual Result: PASS.
 - Blocker / Next Action: None.
+
+## 2026-05-27 13:11:55 CST / producer_system_designer
+- 完成内容: Addressed CI `required-gate` failure. The failure was `check-rust-file-size` reporting `crates/oasis7_client_launcher/src/main_tests.rs` at 1204 lines over the 1200-line test-file limit. Moved the newly added chain network tier args test into `main_tests_chain_args.rs`, reducing `main_tests.rs` to 1186 lines while keeping the behavior coverage.
+- 遗留事项: None for the CI failure.
+- Action: Push follow-up commit and wait for PR required-gate to rerun.
+- Validation Command: `./scripts/check-rust-file-size.sh`
+- Expected Result: No oversized Rust files.
+- Actual Result: PASS, `check-rust-file-size: OK`.
+- Validation Command: `env -u RUSTC_WRAPPER cargo test -p oasis7_client_launcher build_chain_runtime_args_resolves_public_testnet_manifest_from_tier -- --nocapture`
+- Expected Result: Moved chain args test remains discovered and passing.
+- Actual Result: PASS, 1 test passed. Existing unrelated warnings remain in `main_chain_status.rs`.
+- Validation Command: `cargo fmt --check`
+- Expected Result: Rust formatting is current.
+- Actual Result: PASS.
+- Validation Command: `git diff --check`
+- Expected Result: No whitespace errors in the working diff.
+- Actual Result: PASS.
+- Validation Command: `./scripts/pm/lint.sh`
+- Expected Result: `.pm` task and execution log satisfy repo lint.
+- Actual Result: PASS.
+- Blocker / Next Action: None.
