+++ crates/oasis7_client_launcher/src/main_tests.rs
@@ -41,6 +41,8 @@ use std::path::PathBuf;
 use std::sync::{Mutex, OnceLock};
 use std::time::{SystemTime, UNIX_EPOCH};
 
+#[path = "main_tests_chain_args.rs"]
+mod chain_args_tests;
 #[path = "main_tests_explorer.rs"]
 mod explorer_tests;
 #[path = "main_tests_onboarding.rs"]
@@ -387,26 +389,6 @@ fn build_chain_runtime_args_contains_chain_overrides_when_enabled() {
     );
     assert!(args.contains(&"/ip4/127.0.0.1/tcp/4100".to_string()));
     assert!(args.contains(&"/dns4/bootstrap.example/tcp/4101".to_string()));
-}
-
-#[test]
-fn build_chain_runtime_args_resolves_public_testnet_manifest_from_tier() {
-    let mut config = LaunchConfig {
-        chain_enabled: true,
-        chain_status_bind: "127.0.0.1:6121".to_string(),
-        chain_node_id: "chain-a".to_string(),
-        chain_network_tier: "public_testnet".to_string(),
-        chain_p2p_user_mode: "public_entry".to_string(),
-        chain_p2p_accept_public_entry: true,
-        ..LaunchConfig::default()
-    };
-    config.normalize();
-
-    let args = build_chain_runtime_args(&config).expect("args should build");
-
-    assert!(args.contains(&"--network-tier-manifest".to_string()));
-    assert!(args
-        .contains(&"doc/testing/templates/network-tier-public-testnet.example.json".to_string()));
 }
 
 #[test]
